+++ src/components/SectionCard/index.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Meta, Story } from '@storybook/react/types-6-0'
 import { SectionCard } from '@/components/SectionCard'
-import { mockGenEdCourse, mockNoGenEdCourse } from '@/components/CourseCard/mockCourse'
+import { mockGenEdCourse, mockNoGenEdCourse } from '@/modules/CourseSearch/component/CourseCard/mockCourse'
 
 type CourseCardStoryProps = { courseType: 'gened' | 'noGened'; availability: 'available' | 'full' | 'closed' }
