+++ src/modules/CourseSearch/component/CourseCard/components/Caption.tsx
