+++ src/modules/CourseSearch/component/CourseCard/components/CardHeading.tsx
