+++ src/modules/CourseSearch/component/CourseCard/components/SectionSelect.tsx
