+++ src/modules/CourseSearch/component/CourseCard/constants.ts
