+++ src/modules/CourseSearch/component/CourseCard/mockCourse.tsx
