+++ src/modules/CourseSearch/component/CourseCard/utils.ts
