+++ src/modules/CourseSearch/component/CourseList/index.stories.tsx
@@ -1,0 +1,21 @@
+import React from 'react'
+import { CourseList, CourseListProps } from '.'
+import { Meta, Story } from '@storybook/react/types-6-0'
+import { CourseSearchProvider } from '@/modules/CourseSearch/context/CourseSearch'
+
+type StoryProps = CourseListProps
+
+export default {
+  title: 'Component/CourseList',
+  component: CourseList,
+} as Meta<StoryProps>
+
+export const Default: Story<StoryProps> = (args) => {
+  return (
+    <CourseSearchProvider>
+      <CourseList {...args} />
+    </CourseSearchProvider>
+  )
+}
+
+Default.args = {}
