+++ src/modules/CourseSearch/component/FilterIconButton/index.stories.tsx
