+++ src/modules/CourseSearch/component/FilterIconButton/styled.tsx
