+++ src/modules/CourseSearch/component/FilterSection/components/CheckboxGroup/index.stories.tsx
