+++ src/modules/CourseSearch/component/FilterSection/styled.tsx
