+++ src/modules/CourseSearch/component/SearchField/styled.tsx
