+++ src/modules/CourseSearch/component/SelectedCoursesButton/index.tsx
