+++ src/modules/CourseSearch/context/CourseSearch/constants.tsx
@@ -1,0 +1,7 @@
+import { CourseSearchProps } from '@/modules/CourseSearch/context/CourseSearch/types'
+
+export const LIMIT_QUERY_CONSTANT = 15
+
+export const DEFAULT_COURSE_SEARCH_CONTEXT_VALUE: CourseSearchProps = {
+  fetchMoreCourses: () => {},
+}
