+++ src/modules/CourseSearch/hooks/useSearchCourseQueryParams/index.tsx
