+++ src/modules/CourseSearch/types/index.tsx
@@ -1,0 +1,6 @@
+import { SearchCourseResponse, SearchCourseVars } from '@/utils/network/BackendGQLQueries'
+
+export interface CourseSearchPagePrefetchData {
+  vars: SearchCourseVars
+  data: SearchCourseResponse
+}
